requirements.txt
git-filter-repo==2.47.0
python-dotenv==1.2.1
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
schedule>=1.2.0
